Change `CalculateAverageOrderAmountAsync` to `Task<decimal>`
Since an average of decimals is being calculated.

+++ RestaurantReservation.Db/Interfaces/IOrderRepository.cs
@@ -6,5 +6,5 @@ public interface IOrderRepository : IRepository<Order>
 {
   Task<IEnumerable<Order>> ListOrdersAndMenuItemsAsync(int reservationId);
 
-  Task<double> CalculateAverageOrderAmountAsync(int employeeId);
+  Task<decimal> CalculateAverageOrderAmountAsync(int employeeId);
 }
